+++ example/test/golden_ahem_text/ahem_text_demo_test.dart
@@ -31,5 +31,6 @@ void main() {
       ),
     ],
     axes: const MatrixAxes(themes: [MatrixTheme.light, MatrixTheme.dark]),
+    reportFormats: defaultReportFormats,
   );
 }